Refactor background directories
