+++ src/background/components/background.js
@@ -3,7 +3,7 @@ import * as commandActions from 'background/actions/command';
 import * as settingActions from 'background/actions/setting';
 import * as findActions from 'background/actions/find';
 import * as tabActions from 'background/actions/tab';
-import * as commands from 'shared/commands';
+import * as completions from '../shared/completions';
 
 export default class BackgroundComponent {
   constructor(store) {
@@ -44,7 +44,7 @@ export default class BackgroundComponent {
     case messages.SETTINGS_QUERY:
       return Promise.resolve(this.store.getState().setting.value);
     case messages.CONSOLE_QUERY_COMPLETIONS:
-      return commands.complete(message.text, settings.value);
+      return completions.complete(message.text, settings.value);
     case messages.SETTINGS_RELOAD:
       this.store.dispatch(settingActions.load());
       return this.broadcastSettingsChanged();
