+++ src/background/shared/completions/histories.js
