--- src/shared/commands/index.js
@@ -1,3 +1,0 @@
-import complete from './complete';
-
-export { complete };
